// src/models/Simptom.ts
export interface Positive {
  id?: string;
  positiveSign: string;
}
export interface Negative {
  id?: string;
  negativeSign: string;
}
export interface ArmsNegative {
  id?: string;
  negativeSign: string;
}
export interface backNegative {
  id?: string;
  negativeSign: string;
}
export interface globalNegative {
  id?: string;
  negativeSign: string;
}
export interface headNegative {
  id?: string;
  negativeSign: string;
}
export interface InternalNegative {
  id?: string;
  negativeSign: string;
}
export interface LegsNegative {
  id?: string;
  negativeSign: string;
}
export interface NeckNegative {
  id?: string;
  negativeSign: string;
}
export interface SexsualNegative {
  id?: string;
  negativeSign: string;
}
export interface SpiritNegative {
  id?: string;
  negativeSign: string;
}
export interface BodyNegative {
  id?: string;
  negativeSign: string;
}